misc: update minify-trace scripts; extract to lib/ (#8968)
