+++ lighthouse-core/scripts/lantern/minify-devtoolslog.js
@@ -14,6 +14,7 @@
 
 const fs = require('fs');
 const path = require('path');
+const {minifyDevtoolsLog} = require('../../lib/minify-devtoolslog.js');
 
 if (process.argv.length !== 4) {
   console.error('Usage $0: <input file> <output file>');
@@ -26,79 +27,9 @@ const inputDevtoolsLogRaw = fs.readFileSync(inputDevtoolsLogPath, 'utf8');
 /** @type {LH.DevtoolsLog} */
 const inputDevtoolsLog = JSON.parse(inputDevtoolsLogRaw);
 
-const headersToKeep = new Set([
-  // Request headers
-  'accept',
-  'accept-encoding',
-  'accept-ranges',
-  // Response headers
-  'status',
-  'content-length',
-  'content-type',
-  'content-encoding',
-  'content-range',
-  'etag',
-  'cache-control',
-  'last-modified',
-  'link',
-  'x-robots-tag',
-]);
-
-/** @param {Partial<LH.Crdp.Network.Headers>} [headers] */
-function cleanHeaders(headers) {
-  if (!headers) return;
-
-  for (const key of Object.keys(headers)) {
-    if (!headersToKeep.has(key.toLowerCase())) headers[key] = undefined;
-  }
-}
-
-/** @param {{url: string}} obj */
-function cleanDataURI(obj) {
-  obj.url = obj.url.replace(/^(data:.*?base64,).*/, '$1FILLER');
-}
-
-/** @param {LH.Crdp.Network.Response} [response] */
-function cleanResponse(response) {
-  if (!response) return;
-  cleanDataURI(response);
-  cleanHeaders(response.requestHeaders);
-  cleanHeaders(response.headers);
-  response.securityDetails = undefined;
-  response.headersText = undefined;
-  response.requestHeadersText = undefined;
-
-  /** @type {any} */
-  const timing = response.timing || {};
-  for (const [k, v] of Object.entries(timing)) {
-    if (v === -1) timing[k] = undefined;
-  }
-}
-
-/** @param {LH.DevtoolsLog} log */
-function filterDevtoolsLogEvents(log) {
-  return log.map(original => {
-    /** @type {LH.Protocol.RawEventMessage} */
-    const entry = JSON.parse(JSON.stringify(original));
-
-    switch (entry.method) {
-      case 'Network.requestWillBeSent':
-        cleanDataURI(entry.params.request);
-        cleanHeaders(entry.params.request.headers);
-        cleanResponse(entry.params.redirectResponse);
-        break;
-      case 'Network.responseReceived':
-        cleanResponse(entry.params.response);
-        break;
-    }
-
-    return entry;
-  });
-}
-
-const filteredLog = filterDevtoolsLogEvents(inputDevtoolsLog);
+const outputDevtoolsLog = minifyDevtoolsLog(inputDevtoolsLog);
 const output = `[
-${filteredLog.map(e => '  ' + JSON.stringify(e)).join(',\n')}
+${outputDevtoolsLog.map(e => '  ' + JSON.stringify(e)).join(',\n')}
 ]`;
 
 /** @param {string} s */
